Move seq_compare into its own header file
