+++ include/tins/detail/sequence_number_helpers.h
@@ -27,24 +27,23 @@
  *
  */
 
-#include "internals.h"
+#ifndef TINS_SEQUENCE_NUMBER_HELPERS_H
+#define TINS_SEQUENCE_NUMBER_HELPERS_H
 
+#include <stdint.h>
+/**
+ * \cond
+ */
 namespace Tins {
 namespace Internals {
 
-int seq_compare(uint32_t seq1, uint32_t seq2) {
-    // As defined by RFC 1982 - 2 ^ (SERIAL_BITS - 1)
-    static const uint32_t seq_number_diff = 2147483648U;
-    if (seq1 == seq2) {
-        return 0;
-    }
-    if (seq1 < seq2) {
-        return (seq2 - seq1 < seq_number_diff) ? -1 : 1;
-    }
-    else {
-        return (seq1 - seq2 > seq_number_diff) ? -1 : 1;
-    }
-}
+// Compares sequence numbers as defined by RFC 1982.
+int seq_compare(uint32_t seq1, uint32_t seq2);
 
 } // namespace Internals
 } // namespace Tins
+/**
+ * \endcond
+ */
+
+#endif // TINS_SEQUENCE_NUMBER_HELPERS_H
